API sessions are cached by API username
This means that if two modules independently create BarryDonations objects with the same API credentials, they will be given references to the same BarryDonations object. This is good not only because it saves bandwidth, but the current API actually can't support two sessions for a given user, making this functionality mandatory.